# src/FlyingShadow.Api/OpenApi/FlyingShadow.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Flying Daggers API",
    "description": "Flying Daggers API management",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:7113"
    }
  ],
  "paths": {
    "/api/Authentication/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDetails"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDetails"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Authentication/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Battle/start": {
      "post": {
        "tags": [
          "Battle"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BattleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BattleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BattleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BattleResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BattleResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BattleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/FlyingShadow/Shadows": {
      "get": {
        "tags": [
          "FlyingShadow"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShadowDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AcrobaticsLevel": {
        "enum": [
          "Beginner",
          "Intermediate",
          "Advanced"
        ]
      },
      "BattleRequest": {
        "required": [
          "shadowOneCodeName",
          "shadowTwoCodeName"
        ],
        "type": "object",
        "properties": {
          "shadowOneCodeName": {
            "type": "string"
          },
          "shadowTwoCodeName": {
            "type": "string"
          }
        }
      },
      "BattleResponse": {
        "required": [
          "outcome",
          "shadowOneStats",
          "shadowTwoStats",
          "statBreakdown"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "type": "string"
          },
          "shadowOneStats": {
            "$ref": "#/components/schemas/Stats"
          },
          "shadowTwoStats": {
            "$ref": "#/components/schemas/Stats"
          },
          "statBreakdown": {
            "$ref": "#/components/schemas/StatResults"
          }
        }
      },
      "Error": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorCode": {
        "enum": [
          "NotFound",
          "AlreadyExists",
          "UnexpectedError",
          "InvalidCredentials",
          "UnableToRetrieveData",
          "UnableToProcessData"
        ]
      },
      "LoginDetails": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "minLength": 5,
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "required": [
          "tokenDetails"
        ],
        "type": "object",
        "properties": {
          "tokenDetails": {
            "$ref": "#/components/schemas/TokenDetails"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "Rank": {
        "enum": [
          "Danza",
          "Toshiyama",
          "Oniwaban"
        ]
      },
      "RegisterRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "minLength": 5,
            "type": "string"
          }
        }
      },
      "RegisterResponse": {
        "required": [
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ShadowDto": {
        "required": [
          "id",
          "codeName",
          "clan",
          "origin",
          "rank",
          "shadowSkills"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "codeName": {
            "type": "string"
          },
          "clan": {
            "type": "string"
          },
          "origin": {
            "type": "string"
          },
          "rank": {
            "$ref": "#/components/schemas/Rank"
          },
          "shadowSkills": {
            "$ref": "#/components/schemas/StealthMetricsDto"
          }
        }
      },
      "StatResults": {
        "required": [
          "combatPowerWinner",
          "evasionIndexWinner",
          "stealthScoreWinner"
        ],
        "type": "object",
        "properties": {
          "combatPowerWinner": {
            "type": "string"
          },
          "evasionIndexWinner": {
            "type": "string"
          },
          "stealthScoreWinner": {
            "type": "string"
          }
        }
      },
      "Stats": {
        "required": [
          "codeName",
          "combatPower",
          "evasionIndex",
          "stealthScore",
          "overallRating"
        ],
        "type": "object",
        "properties": {
          "codeName": {
            "type": "string"
          },
          "combatPower": {
            "type": "number",
            "format": "double"
          },
          "evasionIndex": {
            "type": "number",
            "format": "double"
          },
          "stealthScore": {
            "type": "number",
            "format": "double"
          },
          "overallRating": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "StealthMetricsDto": {
        "type": "object",
        "properties": {
          "shadowBlendScore": {
            "type": "integer",
            "format": "int32"
          },
          "silenceRating": {
            "type": "integer",
            "format": "int32"
          },
          "invisibilityDurationMs": {
            "type": "integer",
            "format": "int32"
          },
          "acrobaticsLevel": {
            "$ref": "#/components/schemas/AcrobaticsLevel"
          }
        }
      },
      "TokenDetails": {
        "required": [
          "token",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Authentication"
    },
    {
      "name": "Battle"
    },
    {
      "name": "FlyingShadow"
    }
  ]
}